+++ flutter_inappwebview_android/lib/src/container_controller.dart
@@ -94,6 +94,14 @@ class AndroidContainerController extends PlatformContainerController
   }
 
   @override
+  Future<bool> clearContainerData(String containerId) async {
+    final ok = await channel?.invokeMethod<bool>('clearContainerData', {
+      'containerId': containerId,
+    });
+    return ok ?? false;
+  }
+
+  @override
   void dispose() {
     // empty
   }